+++ PostHog/PostHogSessionManager.swift
@@ -20,12 +20,10 @@ import Foundation
         case customSessionId = "Custom session set"
     }
 
-    @objc public static var shared: PostHogSessionManager {
-        DI.main.sessionManager
-    }
+    private let config: PostHogConfig
 
-    // Private initializer to prevent multiple instances
-    override init() {
+    init(config: PostHogConfig) {
+        self.config = config
         super.init()
         registerNotifications()
         registerApplicationSendEvent()
@@ -237,8 +235,11 @@ import Foundation
 
     private func registerApplicationSendEvent() {
         #if os(iOS) || os(tvOS)
-            let applicationEventPublisher = DI.main.applicationEventPublisher
-            applicationEventToken = applicationEventPublisher.onApplicationEvent { [weak self] _, _ in
+            guard config.enableSwizzling else {
+                return
+            }
+
+            applicationEventToken = DI.main.applicationEventPublisher.onApplicationEvent { [weak self] _, _ in
                 // update "last active" session
                 // we want to keep track of the idle time, so we need to maintain a timestamp on the last interactions of the user with the app. UIEvents are a good place to do so since it means that the user is actively interacting with the app (e.g not just noise background activity)
                 self?.queue.async {